Implement uri-unescape that works with utf8-c8 encoding
Perl 6 normalizes all strings to Unicode NFC form. Due to this, using this module
it was not possible to decode this string and leave it unchanged, which would then result in us
not being able to access the file by filename, and potentially many other issues.

For some backstory on my inspiration to do this, see:
https://github.com/MoarVM/MoarVM/issues/525