More docs
Also I deleted Capture here because I believe there's no unambiguous coercion to
Capture for the URI::Query object. At the very least, we would want to define it
so that it takes hash-format into consideration and trying to think what that
would look like gave me a headache so I just dropped it instead.